Update pdb_full_region_curated.txt
Remove manual additions to the curated file that already exist in pdb_full_region table in the database. This is in preparation for changing the format of this file to include the results of cmscan for curated accessions and PDB IDs, so that it is ready to combine with the text file to import to the db during the weekly PDB-Rfam mapping.