Add process to publish to FTP site
Normal workflow publishing won't work for this, as we have to do it as
the datamover, which I don't want to run the main process as. That queue
has limits on the time and memory which we may not be able to stay
within. We may change to using that queue in the future though.

To get around this I added a simple process which copies things to the
FTP directory. This is untested but is the general outline of what we
can do to solve this.